+++ tests/testthat/test-users.R
@@ -13,5 +13,5 @@ test_that("lookup_users works", {
   usr_df <- lookup_users(users)
   expect_s3_class(usr_df, "tbl")
   expect_equal(nrow(usr_df), 6)
-  expect_equal(ncol(usr_df), 91)
+  expect_equal(ncol(usr_df), 20)
 })